pyproject.toml
[project]
name = "ragnar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "marcalph", email = "marcalph@protonmail.com" }
]
requires-python = "==3.10.*"
dependencies = [
    "bs4>=0.0.2",
    "chainlit>=2.5.5",
    "cohere>=5.15.0",
    "davia>=0.1.14",
    "fasttext-langdetect>=1.0.5",
    "frontmatter>=3.0.8",
    "langfuse>=2.60.8",
    "levenshtein>=0.27.1",
    "markdown-it-py>=3.0.0",
    "nltk>=3.9.1",
    "weave>=0.51.54",
    "weaviate>=0.1.2",
    "bm25s>=0.2.13",
    "pystemmer>=3.0.0",
    "scikit-learn>=1.7.0",
    "python-dotenv>=1.1.1",
    "loguru>=0.7.3",
    "httpx>=0.28.1",
    "opentelemetry-instrumentation-cohere>=0.40.14",
    "openai>=1.93.0",
]

[project.scripts]
ragnar = "ragnar:main"

# [build-system]
# requires = ["hatchling"]
# build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]

[tool.hatch.build.wheel]
sources = ["./src/ragnar"]
